requirements.txt
flask==3.1.0
google-generativeai==0.8.3
python-dotenv==1.0.1
requests==2.32.3
flask-cors==5.0.0
numpy==1.26.4
pandas==2.2.0
matplotlib==3.8.2
